Export `FractionalShareStockTrade`
